refactor(ui): Tag - convert to typescript
